feat: production readiness ceiling items 1-4
Item 1 - PHPStan type fixes:
- Fix FileUploadService ZipArchive::getNameIndex() false return (add null check)
- Split realpath+strpos check to properly narrow string type
- Fix md5(json_encode()) false return with (string) cast
- Fix PaystackService date(strtotime()) with (int) cast
- Fix FuelTransactionController fopen false return guard
- Add nullsafe ?-> pattern to phpstan.neon.dist ignoreErrors (SimpleXMLElement
  returns null at runtime despite PHPStan typing it non-nullable)
- Regenerate baseline: 2,446 -> 2,346 unique error entries

Item 2 - Test coverage +18 tests (63 -> 81 passing):
- PaystackWebhookTest (8 tests): HMAC-SHA512 signature verification, tamper
  detection, replay attack prevention, unconfigured secret, all event types
- EmailVerificationTest (6 tests): unverified user blocked, verified user passes,
  notification sent on registration, valid/invalid verification link, interface
  implemented on User model
- TeamDataIsolationTest (4 tests): HasTeamFilters scoping, cross-team route 404,
  per-team machine counts, withoutGlobalScopes bypass

Item 3 - N+1 eager loading:
- All major Livewire components and API controllers already use with()
- preventLazyLoading() (Item 4 from previous commit) now detects any regressions
  at test-time in non-production environments

Item 4 - Infrastructure readiness for scaling:
- Add read-replica support to pgsql config via DB_READ_HOST env var with sticky
  sessions; falls back to DB_HOST when not set (zero breaking change)
- Add dedicated 'integrations' queue connection with 180s timeout and after_commit
- Add [program:mines-integrations-worker] to supervisord config running on
  separate process from default queue workers
- Update default worker command to explicit queue priority list

+++ app/Services/PaystackService.php
@@ -187,7 +187,7 @@ class PaystackService
             }
 
             $nextPaymentDate = isset($subscriptionData['next_payment_date'])
-                ? date('Y-m-d H:i:s', strtotime((string) $subscriptionData['next_payment_date']))
+                ? date('Y-m-d H:i:s', (int) strtotime((string) $subscriptionData['next_payment_date']))
                 : null;
 
             Subscription::updateOrCreate(
@@ -315,7 +315,7 @@ class PaystackService
                     'currency' => strtoupper($invoiceData['currency'] ?? 'ZAR'),
                     'status' => ($invoiceData['paid'] ?? false) ? 'paid' : 'open',
                     'issued_at' => isset($invoiceData['created_at'])
-                        ? date('Y-m-d H:i:s', strtotime((string) $invoiceData['created_at']))
+                        ? date('Y-m-d H:i:s', (int) strtotime((string) $invoiceData['created_at']))
                         : now()->toDateTimeString(),
                     'paid_at' => ($invoiceData['paid'] ?? false) ? now()->toDateTimeString() : null,
                     'pdf_url' => null,
